Calculate payroll

+++ src/main/java/model/TotalHourContract.java
@@ -6,18 +6,18 @@ import java.time.temporal.ChronoUnit;
 
 public class TotalHourContract extends Contract {
 
-    private float totalHours;
-    public TotalHourContract(LocalDate startDate, LocalDate finishDate, float payPerHour, float totalHours) {
+    private double totalHours;
+    public TotalHourContract(LocalDate startDate, LocalDate finishDate, double payPerHour, double totalHours) {
         super(startDate, finishDate, payPerHour);
         this.totalHours = totalHours;
     }
 
     @Override
-    public float calculate(LocalDate date) {
+    public double calculate(LocalDate date) {
         if (date.isAfter(this.getStartDate())) {
             if (date.isBefore(this.getFinishDate())) {
-                long timeDifference = ChronoUnit.DAYS.between(this.getStartDate(), date);
-                long totalDaysContract = ChronoUnit.DAYS.between(this.getStartDate(), this.getFinishDate());
+                double timeDifference = ChronoUnit.DAYS.between(this.getStartDate(), date);
+                double totalDaysContract = ChronoUnit.DAYS.between(this.getStartDate(), this.getFinishDate());
                 return this.getPayPerHour() * ((float) timeDifference / (float) totalDaysContract) * totalHours;
             } else {
                 return this.totalHours * this.getPayPerHour();
